fix compil

+++ src/icrc3_canisters/canisters/icrc3_example/impl/src/utils.rs
@@ -1,6 +1,6 @@
 pub fn trace(msg: &str) {
     unsafe {
-        ic0::debug_print(msg.as_ptr() as i32, msg.len() as i32);
+        ic0::debug_print(msg.as_ptr() as usize, msg.len() as usize);
     }
     ic_cdk::println!("{}", msg);
 }
